Extract products and services screens
