+++ app/packages.tsx
@@ -3,6 +3,8 @@ import React from "react";
 import AuthenticatedApp from "../src/app/AuthenticatedApp";
 import { cashRegisterRenderer } from "./cash-register";
 import { bookingsRenderer } from "./bookings";
+import { productsRenderer } from "./products";
+import { servicesRenderer } from "./services";
 
 export default function Packages(): React.ReactElement {
   return (
@@ -10,6 +12,8 @@ export default function Packages(): React.ReactElement {
       initialScreen="packages"
       renderBookings={bookingsRenderer}
       renderCashRegister={cashRegisterRenderer}
+      renderProducts={productsRenderer}
+      renderServices={servicesRenderer}
     />
   );
 }
